+++ src/lib/js/string.rs
@@ -4,6 +4,7 @@ use crate::{
         function::NativeFunctionData,
         object::{Object, ObjectKind, PROTOTYPE},
         property::Property,
+        regexp::{make_regexp, match_all as regexp_match_all, r#match as regexp_match},
         value::{from_value, to_value, ResultValue, Value, ValueData},
     },
 };
@@ -422,6 +423,15 @@ pub fn last_index_of(this: &Value, args: &[Value], ctx: &mut Interpreter) -> Res
     Ok(to_value(highest_index))
 }
 
+/// Returns an array whose contents is all the results matching the regular expression, if the global (g) flag is present,
+/// in its absence, only the first complete match and its related capturing groups is returned,
+/// otherwise null is returned if no match is found.
+/// <https://tc39.es/ecma262/#sec-string.prototype.match>
+pub fn r#match(this: &Value, args: &[Value], ctx: &mut Interpreter) -> ResultValue {
+    let re = make_regexp(&to_value(Object::default()), &[args[0].clone()], ctx)?.clone();
+    regexp_match(&re, ctx.value_to_rust_string(this), ctx)
+}
+
 /// Abstract method `StringPad`
 /// Performs the actual string padding for padStart/End.
 /// <https://tc39.es/ecma262/#sec-stringpad/>
@@ -682,6 +692,42 @@ pub fn value_of(this: &Value, args: &[Value], ctx: &mut Interpreter) -> ResultVa
     to_string(this, args, ctx)
 }
 
+/// TODO: update this method to return iterator
+/// Returns an array* of all results matching a string against a regular expression, including capturing groups
+/// <https://tc39.es/ecma262/#sec-string.prototype.matchall>
+pub fn match_all(this: &Value, args: &[Value], ctx: &mut Interpreter) -> ResultValue {
+    let re: Value = match args.get(0) {
+        Some(arg) => {
+            if arg == &Gc::new(ValueData::Null) {
+                make_regexp(
+                    &to_value(Object::default()),
+                    &[
+                        to_value(ctx.value_to_rust_string(arg)),
+                        to_value(String::from("g")),
+                    ],
+                    ctx,
+                )
+            } else if arg == &Gc::new(ValueData::Undefined) {
+                make_regexp(
+                    &to_value(Object::default()),
+                    &[Gc::new(ValueData::Undefined), to_value(String::from("g"))],
+                    ctx,
+                )
+            } else {
+                from_value(arg.clone()).map_err(to_value)
+            }
+        }
+        None => make_regexp(
+            &to_value(Object::default()),
+            &[to_value(String::new()), to_value(String::from("g"))],
+            ctx,
+        ),
+    }?
+    .clone();
+
+    regexp_match_all(&re, ctx.value_to_rust_string(this))
+}
+
 /// Create a new `String` object
 pub fn create_constructor(global: &Value) -> Value {
     // Create constructor function object
@@ -709,6 +755,7 @@ pub fn create_constructor(global: &Value) -> Value {
     proto.set_field_slice("includes", to_value(includes as NativeFunctionData));
     proto.set_field_slice("indexOf", to_value(index_of as NativeFunctionData));
     proto.set_field_slice("lastIndexOf", to_value(last_index_of as NativeFunctionData));
+    proto.set_field_slice("match", to_value(r#match as NativeFunctionData));
     proto.set_field_slice("padEnd", to_value(pad_end as NativeFunctionData));
     proto.set_field_slice("padStart", to_value(pad_start as NativeFunctionData));
     proto.set_field_slice("trim", to_value(trim as NativeFunctionData));
@@ -718,6 +765,7 @@ pub fn create_constructor(global: &Value) -> Value {
     proto.set_field_slice("substring", to_value(substring as NativeFunctionData));
     proto.set_field_slice("substr", to_value(substr as NativeFunctionData));
     proto.set_field_slice("valueOf", to_value(value_of as NativeFunctionData));
+    proto.set_field_slice("matchAll", to_value(match_all as NativeFunctionData));
 
     let string = to_value(string_constructor);
     proto.set_field_slice("constructor", string.clone());
@@ -878,5 +926,119 @@ mod tests {
         assert_eq!(forward(&mut engine, "emptyLiteral.endsWith('')"), pass);
         assert_eq!(forward(&mut engine, "enLiteral.endsWith('h')"), pass);
         assert_eq!(forward(&mut engine, "zhLiteral.endsWith('文')"), pass);
+    }
+
+    #[test]
+    fn match_all() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+
+        assert_eq!(
+            forward(&mut engine, "'aa'.matchAll(null).length"),
+            String::from("0")
+        );
+        assert_eq!(
+            forward(&mut engine, "'aa'.matchAll(/b/).length"),
+            String::from("0")
+        );
+        assert_eq!(
+            forward(&mut engine, "'aa'.matchAll(/a/).length"),
+            String::from("1")
+        );
+        assert_eq!(
+            forward(&mut engine, "'aa'.matchAll(/a/g).length"),
+            String::from("2")
+        );
+
+        forward(
+            &mut engine,
+            "const groupMatches = 'test1test2'.matchAll(/t(e)(st(\\d?))/g)",
+        );
+        assert_eq!(
+            forward(&mut engine, "groupMatches.length"),
+            String::from("2")
+        );
+        assert_eq!(
+            forward(&mut engine, "groupMatches[0][1]"),
+            String::from("e")
+        );
+        assert_eq!(
+            forward(&mut engine, "groupMatches[0][2]"),
+            String::from("st1")
+        );
+        assert_eq!(
+            forward(&mut engine, "groupMatches[0][3]"),
+            String::from("1")
+        );
+        assert_eq!(
+            forward(&mut engine, "groupMatches[1][3]"),
+            String::from("2")
+        );
+
+        assert_eq!(
+            forward(
+                &mut engine,
+                "'test1test2'.matchAll(/t(e)(st(\\d?))/).length"
+            ),
+            String::from("1")
+        );
+
+        let init = r#"
+        const regexp = RegExp('foo[a-z]*','g');
+        const str = 'table football, foosball';
+        const matches = str.matchAll(regexp);
+        "#;
+        forward(&mut engine, init);
+        assert_eq!(
+            forward(&mut engine, "matches[0][0]"),
+            String::from("football")
+        );
+        assert_eq!(forward(&mut engine, "matches[0].index"), String::from("6"));
+        assert_eq!(
+            forward(&mut engine, "matches[1][0]"),
+            String::from("foosball")
+        );
+        assert_eq!(forward(&mut engine, "matches[1].index"), String::from("16"));
+    }
+
+    #[test]
+    fn test_match() {
+        let realm = Realm::create();
+        let mut engine = Executor::new(realm);
+        let init = r#"
+        var str = new String('The Quick Brown Fox Jumps Over The Lazy Dog');
+        var result1 = str.match(/quick\s(brown).+?(jumps)/i);
+        var result2 = str.match(/[A-Z]/g);
+        var result3 = str.match("T");
+        var result4 = str.match(RegExp("B", 'g'));
+        "#;
+
+        forward(&mut engine, init);
+        assert_eq!(forward(&mut engine, "result1[0]"), "Quick Brown Fox Jumps");
+        assert_eq!(forward(&mut engine, "result1[1]"), "Brown");
+        assert_eq!(forward(&mut engine, "result1[2]"), "Jumps");
+        assert_eq!(forward(&mut engine, "result1.index"), "4");
+        assert_eq!(
+            forward(&mut engine, "result1.input"),
+            "The Quick Brown Fox Jumps Over The Lazy Dog"
+        );
+
+        assert_eq!(forward(&mut engine, "result2[0]"), "T");
+        assert_eq!(forward(&mut engine, "result2[1]"), "Q");
+        assert_eq!(forward(&mut engine, "result2[2]"), "B");
+        assert_eq!(forward(&mut engine, "result2[3]"), "F");
+        assert_eq!(forward(&mut engine, "result2[4]"), "J");
+        assert_eq!(forward(&mut engine, "result2[5]"), "O");
+        assert_eq!(forward(&mut engine, "result2[6]"), "T");
+        assert_eq!(forward(&mut engine, "result2[7]"), "L");
+        assert_eq!(forward(&mut engine, "result2[8]"), "D");
+
+        assert_eq!(forward(&mut engine, "result3[0]"), "T");
+        assert_eq!(forward(&mut engine, "result3.index"), "0");
+        assert_eq!(
+            forward(&mut engine, "result3.input"),
+            "The Quick Brown Fox Jumps Over The Lazy Dog"
+        );
+        assert_eq!(forward(&mut engine, "result4[0]"), "B");
     }
 }